reader_api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
ada-url==1.27.0
    # via auth0-api-python
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via litellm
aiosignal==1.4.0
    # via aiohttp
alembic==1.17.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
asgiref==3.10.0
    # via opentelemetry-instrumentation-asgi
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
auth0-api-python==1.0.0b6
    # via auth0-fastapi-api
auth0-fastapi-api==1.0.0b4
    # via -r requirements.in
authlib==1.6.5
    # via auth0-api-python
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   ada-url
    #   cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   litellm
    #   typer-slim
    #   uvicorn
cryptography==46.0.3
    # via authlib
distro==1.9.0
    # via openai
dotenv==0.9.9
    # via -r requirements.in
fastapi==0.115.14
    # via
    #   -r requirements.in
    #   auth0-fastapi-api
fastuuid==0.14.0
    # via litellm
filelock==3.20.0
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via huggingface-hub
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   auth0-api-python
    #   auth0-fastapi-api
    #   huggingface-hub
    #   litellm
    #   openai
huggingface-hub==1.1.2
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via
    #   litellm
    #   opentelemetry-api
jinja2==3.1.6
    # via litellm
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
litellm==1.79.3
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
openai==2.7.2
    # via litellm
opentelemetry-api==1.38.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-asyncio
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-distro==0.59b0
    # via -r requirements.in
opentelemetry-exporter-otlp==1.38.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.38.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.38.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.59b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-asyncio
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-logging
opentelemetry-instrumentation-asgi==0.59b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asyncio==0.59b0
    # via -r requirements.in
opentelemetry-instrumentation-fastapi==0.59b0
    # via -r requirements.in
opentelemetry-instrumentation-httpx==0.59b0
    # via -r requirements.in
opentelemetry-instrumentation-logging==0.59b0
    # via -r requirements.in
opentelemetry-proto==1.38.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.38.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.59b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-asyncio
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.59b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
packaging==25.0
    # via
    #   huggingface-hub
    #   opentelemetry-instrumentation
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.0
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   litellm
    #   openai
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via
    #   dotenv
    #   litellm
    #   uvicorn
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   auth0-api-python
    #   opentelemetry-exporter-otlp-proto-http
    #   tiktoken
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via huggingface-hub
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.27
    # via -r requirements.in
sse-starlette==3.0.3
    # via -r requirements.in
starlette==0.46.2
    # via fastapi
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.1
    # via litellm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   grpcio
    #   huggingface-hub
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.34.3
    # via
    #   -r requirements.in
    #   auth0-fastapi-api
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asyncio
    #   opentelemetry-instrumentation-httpx
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
